% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{calc_es50}
\alias{calc_es50}
\title{Calculate ES50 (Hurlbert index)}
\usage{
calc_es50(df, esn = 50)
}
\arguments{
\item{df}{data frame}

\item{esn}{expected number of marine species}
}
\value{
value
}
\description{
Calculate the expected number of marine species in a random sample of 50 individuals (records)
}
\details{
The expected number of marine species in a random sample of 50 individuals (records) is an indicator on marine biodiversity richness.
The ES50 is defined in OBIS as the \code{sum(esi)} over all species of the following per species calculation:
\itemize{
\item when `n - ni >= 50 (with n as the total number of records in the cell and ni the total number of records for the ith-species)
\itemize{
\item \code{esi = 1 - exp(lngamma(n-ni+1) + lngamma(n-50+1) - lngamma(n-ni-50+1) - lngamma(n+1))}
}
\item when \code{n >= 50}
\itemize{
\item \code{esi = 1}
}
\item else
\itemize{
\item \code{esi = NULL}
Warning: ES50 assumes that individuals are randomly distributed, the sample size is sufficiently large, the samples are taxonomically similar, and that all of the samples have been taken in the same manner.
}
}
}
\concept{analyze}
