% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{calc_indicators}
\alias{calc_indicators}
\title{Calculate Biodiversity Indicators, including ES50 (Hurlbert index)}
\usage{
calc_indicators(df, esn = 50)
}
\arguments{
\item{df}{data frame with unique species observations containing columns:
\code{cell}, \code{species}, \code{records}}

\item{esn}{expected number of marine species}
}
\value{
Data frame with the following extra columns: - \code{n}: number of records
\itemize{
\item \code{sp}: species richness - \code{shannon}: Shannon index - \code{simpson}: Simpson
index - \code{es}: Hurlbert index (n = 50), i.e. expected species from 50
samples ES(50) - \code{hill_1}: Hill number \code{exp(shannon)} - \code{hill_2}: Hill
number \code{1/simpson} - \code{hill_inf}: Hill number \code{1/maxp}
}
}
\description{
Calculate the expected number of marine species in a random sample of 50
individuals (records)
}
\details{
The expected number of marine species in a random sample of 50
individuals (records) is an indicator on marine biodiversity richness. The
ES50 is defined in OBIS as the \code{sum(esi)} over all species of the following
per species calculation:
\itemize{
\item when `n - ni >= 50 (with n as the total number of records in the cell and
ni the total number of records for the ith-species)
\itemize{
\item \code{esi = 1 - exp(lngamma(n-ni+1) + lngamma(n-50+1) - lngamma(n-ni-50+1) - lngamma(n+1))}
}
\item when \code{n >= 50} - \code{esi = 1}
\item else - \code{esi = NULL}
}

Warning: ES50 assumes that individuals are randomly distributed, the sample
size is sufficiently large, the samples are taxonomically similar, and that
all of the samples have been taken in the same manner.
}
\concept{analyze}
