% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3.R
\name{make_hex_res}
\alias{make_hex_res}
\title{Make hexagon feature}
\usage{
make_hex_res(hex_res = 2)
}
\arguments{
\item{hex_res}{resolution of H3 hexagons;
see \href{https://h3geo.org/docs/core-library/restable/}{Table of Cell Areas for H3 Resolutions | H3}}
}
\value{
spatial feature \code{sf} object
}
\description{
Make hexagon feature
}
\examples{
   hexes <- make_hex_res(0)
}
\concept{visualize}
