% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eov.R
\name{obis_eov_aphiaid}
\alias{obis_eov_aphiaid}
\title{AphiaID seeds for one or more EOVs}
\usage{
obis_eov_aphiaid(eov)
}
\arguments{
\item{eov}{EOV name(s); see \code{\link[=obis_eov_seeds]{obis_eov_seeds()}}.}
}
\value{
integer vector of seed AphiaIDs (union across the named EOVs).
}
\description{
AphiaID seeds for one or more EOVs
}
\concept{eov}
