% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occ_SAtlantic}
\alias{occ_SAtlantic}
\title{OBIS occurrences, South Atlantic full regional sample}
\format{
A data frame with 1,014,006 rows and 4 variables:
\itemize{
\item \code{decimalLongitude}: unique x
\item \code{decimalLatitude}: unique y
\item \code{species}: unique species
\item \code{records}: number of records based on unique values of x, y, species
}
}
\source{
\url{https://obis.org/data/access}
}
\usage{
occ_SAtlantic
}
\description{
Simple dataset of species and location for example purposes from 2022-04-04.
}
\concept{data}
\keyword{datasets}
